Initial implementation of Go PoP token builder/validator
This commit contains the initial implementation of the Go PoP token
builder and validator package.  The package provides a PoPToken type, a
New function to create it, and several options that may be provided to New
to influence how the token is created.  The PoPToken type provides a
Build method, to compute a PoP token from an HTTP request; a Sign method,
to actually attach the computed PoP token to the HTTP request in the form
of an "X-Authorization" header; and a Verify method, to verify that an
HTTP request contains an "X-Authorization" header and to verify that it
matches the rest of the request.

Although the code is clean, according to the lint tests, this commit does
not include any unit tests as of yet.  Additionally, the handling of
repeated headers is not completely clear; the chosen implementation is
based on RFC 9110's description of how repeated headers is to be handled,
but it is not yet clear that this is the correct implementation of this
case.  As such, the logic has been centralized into a single utility
function, to allow easy refactoring should this not be the correct
algorithm.