Correct a typo in go.mod
A minor typo slipped into the go.mod in the code I recently contributed.
It didn't show up until I attempted to use the module.  I had tried to use
it previously, but misinterpreted the errors I recieved as being due to
the fact that the code did not yet exist in its canonical location.  This
commit corrects the typo.